+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.2]
+
+- Add support for MaxMind (free) license key requirement to update GeoLite2 databases
+- Switch to [ghru](https://github.com/axllent/ghru) for binary release updates
+
+
 ## [0.2.1]
 
 - Fix dataDir flag parsing